feat: file watch
