+++ extensions/iceworks-time-master/src/utils/sender.ts
@@ -3,8 +3,8 @@ import * as path from 'path';
 import axios from 'axios';
 import * as fse from 'fs-extra';
 import { ALI_DIP_PRO } from '@iceworks/constant';
-import { KeystrokeStats } from '../recorders/keystrokeStats';
-import { FileChange, FileChangeInfo, FileEventInfo } from '../storages/filesChange';
+import { KeystrokeStats, FileChange } from '../recorders/keystrokeStats';
+import { FileChangeInfo, FileEventInfo } from '../storages/file';
 import { getStoragePayloadsPath } from './storage';
 import { getEditorInfo, getExtensionInfo, getSystemInfo, SystemInfo, EditorInfo, ExtensionInfo } from './env';
 import { ProjectInfo } from '../storages/project';
